mac80211: Replace ath9k dynack with static distance setting
The ath9k dynack setting seems to have caused tx path hangs. Especially in
crowded environments. So instead define a high distance setting to make
sure that it still works over large links (up to 3km). There is a reduction
in performance for installations in a confined space - but we have to life
with it.

This not only affects ath9k but also other drivers which support static
distance settings.

Signed-off-by: Sven Eckelmann <sven@narfation.org>